Setup environment
