+++ package.json
@@ -20,6 +20,7 @@
   },
   "homepage": "https://github.com/cuauh-cabrera/backend-project-bedu#readme",
   "dependencies": {
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "mysql2": "^3.6.2"
   },
